requirements.txt
flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
python-dotenv==1.0.1
google-generativeai==0.8.3
openai==1.51.2
httpx==0.27.2
sounddevice==0.4.7
soundfile==0.12.1
numpy==2.1.2
playsound==1.3.0
faster-whisper==1.0.3
librosa==0.10.1
